tests.www.apply: Split, cleanup and pytest-style
